requirements.txt
streamlit
sentence-transformers
faiss-cpu
transformers
pypdf
torch
accelerate
ollama